best.R functioning correctly
After testing best.R appears to be functioning correctly using the
examples in the assignment brief.

Troubleshooting involved problems using the value of a variable to
subset a list using the $ sign. This didn't work as the $ sign can only
be used with literal names and not computed indices. In this case the [[
]] notation must be used.

Also, initially I had forgotten to convert the data in the required
column from characters to numerics.