Assignment 3 Part 2 - rankhospital.R
For the second part of the assignment, we needed to create a script
'rankhospital.R' that would return the hospital at a specified ranking
position.
This initial commit of the 'rankhospital.R' script is functioning
correctly according to the example tests in the assignment brief. If all
tests pass during submission, then there will be no amendments necessary
and this will be the finished/submitted version.