Assignment 3 Part 3 - rankall.R
The rankall.R script returns a data frame of hospitals in all states
that correspond to the specified feature and ranking.
The 'rankall.R' script is functioning correctly with all the example
tests in the assignment brief. If all tests are passed during
submission, then no amendments will be necessary and this commit will be
the finished version.

Note: This version of 'rankall.R' is an iterative implementation. I
would like to see if I can re-implement it using a vectorized approach,
such as with 'lapply' or 'sapply'.