Fixed merging test & training data set
On the surface, merging of the two data sets seems to be functioning
correctly indicated by the fact that all columns/variables are present.
Use of 'rbindfill' function.
Strange that 'rbind' with a list of data sets would not correctly copy
the columns. Something would become dis-ordered at column num 318 and
only 479 columns would be in the combined data set.