Calculating column means and some tidy up UNTESTED
Implementated step 5 in the project, to calculate the means of each
feature by activity for each participant. The results from this were
stored in a new data table 'mean_data'.

Also some general tidy up of the code, removing obsolete lines and
re-ordering the columns in the full tidy data set so that activity name
came after participant id, as initially this was put at the end when it
was created. This made the table more human readable when printed.

NOTE:
Non of this code has been tested on the full data set, only on a sub-set
of 200 records (100 from the training set & 100 from the test set).