Completed work on Ex Data project 2
Completed working on code for project 2 of Experimental Data Analysis.

This was code to answer questions 3 to 6 of the description using the
ggplot2 package.
This actually took a lot longer then I had expected and involved some
basic data manipulation:
filtering data based on value matching
sub-strings and regular expressions
joining of tables
Hopefully answers are correct as the questions weren't very specific on
what was required or qualified for certain categories.

I should read through the forums and check against grading rubrik to see
if I have misunderstood anything. To be done at a later date.