Assignment 2 Initial Commit
Start of rep data assignment 2.
Commit includes file covering the project description and supplementary
information.
noaa_analysis.Rmd is the main focus of this project and will contain
RMarkdown code, performing the analysis. In this commit the code only
downloads the raw data file and reads the data into R. It also includes
the introductory parts of the analysis.