Initial Analysis
start of data analysis towards answering Q1 of the assignment.
The data has been split into a subset of variable of interest.
Stuck when programming code to extract events with tag "Summary"
indicating that these are summaries of particular events. I believe
these should not be part of the analysis, only the explicit meteological
events.
Need to look into regular expressions to find matching records.