Data Processing Q1
Continued work on data processing to answer question 1.
The data has been processed to discover the most 'harmful' types of
events (excluding 'summary' records).
Incomplete work on presenting the results to show the top 10 most
harmful event types, but maintaining the proportions of fatalities to
injuries.  Work is needed to change the result format from wide to long
to improve suitability for plotting.