Plot Q1
Continued processing of data for question 1.
Plot created of top 10 most harmful types of events in the form of a
stacked bar plot using ggplot2.
The reshaping of the data set from wide to long format was done using
'gather' function in the 'tidyr' package. This was the first time I've
ever used this function.

NOTE: No conclusions have been written based on this plot.
Another plot to show the relative proportion of fatalities to injuries
might be useful in order to show which is more frequent. However on 3
plots are permitted in total, so it might be work looking into what
plots are required for question 2.