Data Processing Q2
Work on data processing for question 2, economic consequences of severe
weather. I decided to use the data.table package for this instead of
dplyr (as in Q1) because I find it very efficient and powerful to do a
lot of processing in a single line of code. Not as clear to follow as
the flow of dplyr.
This was done by computing on the PROPDMG & CROPDMG variables.
I decided to compute the total destruction for each type of event over
two different time scales. The whole data set (62 years) and also by
decade.
Turning the date values which are continuous in to discrete 'decade'
values was interesting as it required creating factors and computing the
level boundaries. This is something I still need more practise with.

Next will be to create a plot of 2 of these results and comment on the
findings. I would like the plot to be a line graph of destructive cost
over decade, faceted by event type..