Tree Plot
Used the Rattle package to create a fancy Rpart plot of the regression
tree used.
Successfully this output to the desired tab.
Had to do this with a function external to the shiny server function.
Should really watch more stuff on reactive functions to understand why
this was necessary.

Next, to implement a scatter plot of results. However the options to
choose the axes should only be for predictors used in the model and also
catch the exception when only 1 predictor was used.