Plotting Problems
Attempted to create a quick scatter plot using ggplot2, but it only
plots a single point (the last point?) and mis-labels the axes.
Previous error said it couldn't find xVar1, which I think it was looking
for in the data and not the local variable with a string assigned to it.