Rattle Removal
Had to remove the rattle package and the fancyRpartPlot function because
rattle is not supported by shinyapps.io and was the reason for my errors
when deploying my app.

Additions:
Intro tab to provide instructions.
numeric input to allow the user to provide the seed number.
I would have liked the numeric input to only appear when required, but a
conditional panel within a conditional panel broke the conditional
panels completely.